// apps/web/src/api/generated/public-tournaments/publicTournamentsControllerListPublicTournamentsMode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type PublicTournamentsControllerListPublicTournamentsMode = typeof PublicTournamentsControllerListPublicTournamentsMode[keyof typeof PublicTournamentsControllerListPublicTournamentsMode];


export const PublicTournamentsControllerListPublicTournamentsMode = {
  ONLINE: 'ONLINE',
  ONSITE: 'ONSITE',
} as const;