// apps/web/src/api/generated/public-tournaments/publicTournamentsControllerListPublicTournamentsSortBy.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type PublicTournamentsControllerListPublicTournamentsSortBy = typeof PublicTournamentsControllerListPublicTournamentsSortBy[keyof typeof PublicTournamentsControllerListPublicTournamentsSortBy];


export const PublicTournamentsControllerListPublicTournamentsSortBy = {
  publishedAt: 'publishedAt',
  startsAt: 'startsAt',
  registrationClosesAt: 'registrationClosesAt',
  name: 'name',
} as const;