// apps/web/src/api/generated/public-tournaments/publicTournamentsControllerListPublicTournamentsSortDirection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type PublicTournamentsControllerListPublicTournamentsSortDirection = typeof PublicTournamentsControllerListPublicTournamentsSortDirection[keyof typeof PublicTournamentsControllerListPublicTournamentsSortDirection];


export const PublicTournamentsControllerListPublicTournamentsSortDirection = {
  asc: 'asc',
  desc: 'desc',
} as const;