// apps/web/src/api/generated/public-tournaments/publicTournamentsControllerListPublicTournamentsStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type PublicTournamentsControllerListPublicTournamentsStatus = typeof PublicTournamentsControllerListPublicTournamentsStatus[keyof typeof PublicTournamentsControllerListPublicTournamentsStatus];


export const PublicTournamentsControllerListPublicTournamentsStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  REGISTRATION_OPEN: 'REGISTRATION_OPEN',
  REGISTRATION_CLOSED: 'REGISTRATION_CLOSED',
  CHECK_IN_OPEN: 'CHECK_IN_OPEN',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  POSTPONED: 'POSTPONED',
  CANCELLED: 'CANCELLED',
  ARCHIVED: 'ARCHIVED',
} as const;